Add files via upload
This project is an implementation of a Credit Card Vault system, designed to securely manage customer information, credit card details, and financial transactions. The system implements role-based access control with different interfaces for administrators, finance users, support users, and customers. It uses encryption (AES) for sensitive data like credit card numbers and hashing (SHA-256) for passwords. The project includes features such as customer registration, transaction management, and a customer dashboard.
XAMPP was used to create and manage the database for this project. XAMPP provides a local Apache web server, MySQL database, and PHP environment, making it ideal for developing and testing web applications. The MySQL database included with XAMPP was used to create and manage the tables for the Credit Card Vault system.